feat(model): add editable PCB item mutation layer
